Fixes and Enhancements:
Resolved issue with email message template in Authentication API for signup.
Completed the signup function in the Authentication service for handling API calls from the frontend.
Created a Signup DTO model to send specific data fields required by the API from the frontend.
Finalized and integrated the Signup page in the Authentication module on the frontend; successfully connected with the signup API and tested functionality.
Applied small bug fixes on the Sign-In page.
Updated the error page for better user experience and error handling.